Allow empty blocks.
When using the markdown code block embedded in the PR body to supply a
changelog, we were discarding any blocks that were empty.

This prevents the use of a `release-note:none` empty block, which is
useful for signaling that you're intentionally not including a release
note for a PR, as it has no user-facing change.

This PR updates to add support for that, and a test to ensure it works
correctly.